+++ adi/devgen/helpers/debug.py
@@ -1,0 +1,3 @@
+def breakpoint():
+    import pdb; pdb.set_trace()
+